Make fNBins a size_t in BinAxis

+++ src/histogram/BinAxis.h
@@ -48,7 +48,7 @@ class BinAxis{
 	bool operator!=(const BinAxis& other_) const;
 
  private:
-    int    fNBins;
+    size_t fNBins;
     double fMin;
     double fMax;
